Autoclaim-V3/autoclaim_project/server/requirements.txt
# AutoClaim Server — Python Dependencies
# Generated from venv on 2026-03-12
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────
fastapi==0.129.0
uvicorn==0.40.0
python-multipart==0.0.22
starlette==0.52.1

# ── Database ──────────────────────────────────────────────────
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
greenlet==3.3.1

# ── Auth / Security ───────────────────────────────────────────
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.5
PyJWT==2.11.0

# ── Data Validation ───────────────────────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
email-validator==2.3.0

# ── AI / ML ───────────────────────────────────────────────────
torch==2.7.1
torchvision==0.22.1
ultralytics==8.4.14
ultralytics-thop==2.0.18
easyocr==1.7.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
numpy==2.4.2
scipy==1.17.0
scikit-image==0.26.0
pillow==12.1.1

# ── Gemini AI (primary OCR) ────────────────────────────────────
google-generativeai>=0.8.0

# ── Groq AI Client ────────────────────────────────────────────
groq==1.0.0

# ── PDF Report Generation ─────────────────────────────────────
Jinja2==3.1.6
MarkupSafe==3.0.3
reportlab==4.4.10

# ── Image Utilities ───────────────────────────────────────────
ImageHash==4.3.2
ImageIO==2.37.2
PyWavelets==1.9.0
tifffile==2026.1.28

# ── Geolocation ───────────────────────────────────────────────
geopy==2.4.1
geographiclib==2.1

# ── HTTP / Networking ─────────────────────────────────────────
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
anyio==4.12.1
sniffio==1.3.1
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3

# ── HuggingFace ───────────────────────────────────────────────
huggingface_hub==1.4.1
hf-xet==1.2.0
filelock==3.23.0
fsspec==2026.2.0

# ── Config / Env ──────────────────────────────────────────────
python-dotenv==1.2.1
PyYAML==6.0.3
click==8.3.1
colorama==0.4.6

# ── Utilities ─────────────────────────────────────────────────
tqdm==4.67.3
psutil==7.2.2
six==1.17.0
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
packaging==26.0
polars==1.38.1
rich==14.3.2

# ── Misc / Transitive ─────────────────────────────────────────
mpmath>=1.3.0
sympy>=1.12
shapely==2.1.2
contourpy==1.3.3
fonttools==4.61.1
matplotlib==3.10.8
cffi==2.0.0
pycparser==3.0
distro==1.9.0
python-bidi==0.6.7
pyclipper==1.4.0
